pyproject.toml
[tool.poetry]
name = "connector_nats"
version = "1.0.0"
description = "Publish messages to NATS through a SpiffWorkflow Service Task"
authors = ["Pit Wegner <pit.wegner@gmail.com>"]
readme = "README.md"
packages = [{include = "connector_nats", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
spiffworkflow-connector-command = {git = "https://github.com/sartography/spiffworkflow-connector-command.git", rev = "main"}
nats-py = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
mypy = "^1.6.0"
ruff = "^0.0.292"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = [
  "B",   # flake8-bugbear
  "C",   # mccabe
  "E",   # pycodestyle error
  # "ERA",  # eradicate
  "F",   # pyflakes
  "N",   # pep8-naming
  "PL",  # pylint
  "S",   # flake8-bandit
  "UP",  # pyupgrade
  "W",   # pycodestyle warning
  "I001" # isort
]

ignore = [
  "C901",  # "complexity" category
  "PLR",  # "refactoring" category has "too many lines in method" type stuff
  "PLE1205"  # saw this Too many arguments for `logging` format string give a false positive once
]

line-length = 130

# target python 3.10
target-version = "py310"

exclude = [
  "migrations"
]

[tool.ruff.per-file-ignores]
"migrations/versions/*.py" = ["E501"]
"tests/**/*.py" = ["PLR2004", "S101"]  # PLR2004 is about magic vars, S101 allows assert

[tool.ruff.isort]
force-single-line = true

[tool.mypy]
strict = true
disallow_any_generics = false
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
